# Make fence in Serial non blocking when Kokkos_ENABLE_ATOMICS_BYPASS is set