@PaulGannay 2025-03-03
I added the `Kokkos::Tools::profileLibraryLoaded` test with the pr #7817 in mind, which makes this function also lightweight, so if pr#7817 is merged, this change shouldn't have an impact on the performances.
Otherwise you are right that I should remove the call to `profileLibraryLoaded`.

Anyways, I'm not fully convinced either that it is needed so if someone else feels your way I will remove it.