@JBludau 2025-03-13
@PaulGannay can you add some documentation for this for the upcoming release, please?